feat(DAO): repository 로직 이관, 변경사항 적용
