+++ src/main/java/janggi/Application.java
@@ -4,7 +4,9 @@ import janggi.controller.GameLobbyController;
 import janggi.domain.game.GameManager;
 import janggi.persistence.DatabaseInitializer;
 import janggi.persistence.DatabaseProvider;
-import janggi.persistence.JDBCGameRepository;
+import janggi.persistence.dao.JdbcBoardDao;
+import janggi.persistence.dao.JdbcGameDao;
+import janggi.persistence.repository.JdbcGameRepository;
 import janggi.service.JanggiService;
 import janggi.view.InputView;
 import janggi.view.OutputView;
@@ -16,7 +18,7 @@ public class Application {
     public static void main(String[] args) {
         InputView inputView = new InputView();
         OutputView outputView = new OutputView();
-        JanggiService janggiService = new JanggiService(new JDBCGameRepository());
+        JanggiService janggiService = new JanggiService(new JdbcGameRepository(new JdbcGameDao(), new JdbcBoardDao()));
         runGameLifecycle(inputView, outputView, janggiService);
         inputView.close();
     }
