+++ src/main/java/janggi/persistence/dao/BoardDao.java
@@ -1,4 +1,14 @@
 package janggi.persistence.dao;
 
-public class BoardDAO {
+import janggi.domain.board.Board;
+import java.sql.Connection;
+import java.sql.SQLException;
+
+public interface BoardDao {
+
+    void deleteByGameId(Connection connection, long gameId) throws SQLException;
+
+    void insertAll(Connection connection, long gameId, Board board) throws SQLException;
+
+    Board findByGameId(Connection connection, long gameId) throws SQLException;
 }
