+++ src/main/java/janggi/persistence/dao/GameDao.java
@@ -1,4 +1,18 @@
 package janggi.persistence.dao;
 
-public class GameDAO {
+import janggi.domain.game.GameManager;
+import janggi.dto.GameSessionDTO;
+import java.sql.Connection;
+import java.sql.SQLException;
+import java.util.List;
+
+public interface GameDao {
+
+    long insert(Connection connection, GameManager gameManager) throws SQLException;
+
+    void update(Connection connection, GameManager gameManager) throws SQLException;
+
+    GameSessionDTO findById(Connection connection, long gameId) throws SQLException;
+
+    List<GameSessionDTO> findAllActive(Connection connection) throws SQLException;
 }
